moves components into separate files
