+++ src/components/Customer.js
@@ -1,0 +1,13 @@
+import React, { useEffect, useState, useCallback } from 'react';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link,
+  useRouteMatch,
+  useParams
+} from "react-router-dom";
+// import PropTypes from 'prop-types';
+// import axios from 'axios';
+
+// import './Customer.css';
