+++ src/components/Home.js
@@ -1,0 +1,11 @@
+import React, { useEffect, useState, useCallback } from 'react';
+// import PropTypes from 'prop-types';
+// import axios from 'axios';
+
+// import './Home.css';
+
+const Home = () => {
+  return <h2>Home</h2>;
+}
+
+export default Home;
